1.0.53.0

+++ SQLite.Designer/Design/Table.cs
@@ -1,3 +1,10 @@
+/********************************************************
+ * ADO.NET 2.0 Data Provider for SQLite Version 3.X
+ * Written by Robert Simpson (robert@blackcastlesoft.com)
+ * 
+ * Released to the public domain, use at your own risk!
+ ********************************************************/
+
 namespace SQLite.Designer.Design
 {
   using System;
@@ -10,7 +17,16 @@ namespace SQLite.Designer.Design
   using System.Text;
   using SQLite.Designer.Editors;
 
-  internal class Table : ICustomTypeDescriptor, IHaveConnection
+  internal abstract class ViewTableBase: IHaveConnection
+  {
+    public abstract string Name { get; set; }
+    public abstract string Catalog { get; }
+    public abstract object Triggers { get; }
+    public abstract void MakeDirty();
+    public abstract DbConnection GetConnection();
+  }
+
+  internal class Table : ViewTableBase, ICustomTypeDescriptor
   {
     private string _name;
     private string _oldname;
@@ -23,6 +39,9 @@ namespace SQLite.Designer.Design
     private List<Index> _oldindexes = new List<Index>();
     private List<ForeignKey> _fkeys = new List<ForeignKey>();
     private List<ForeignKey> _oldfkeys = new List<ForeignKey>();
+    private List<string> _check = new List<string>();
+    private List<Trigger> _triggers = new List<Trigger>();
+    private List<Trigger> _oldtriggers = new List<Trigger>();
     private PrimaryKey _key;
     internal TableDesignerDoc _owner;
     internal DbConnection _connection;
@@ -59,6 +78,11 @@ namespace SQLite.Designer.Design
           _columns.Add(new Column(row, this));
         }
       }
+    }
+
+    public override void MakeDirty()
+    {
+      _owner.MakeDirty();
     }
 
     private void ReloadDefinition()
@@ -100,12 +124,87 @@ namespace SQLite.Designer.Design
         }
       }
 
+      _check.Clear();
       StringBuilder builder = new StringBuilder();
+      SimpleTokenizer.StringParts[] arr = SimpleTokenizer.BreakString(_origSql);
+      for (int n = 0; n < arr.Length - 3; n++)
+      {
+        if (arr[n].keyword == "CONSTRAINT")
+        {
+          builder.Length = 0;
+          int x;
+          for (x = 1; x < 3; x++)
+          {
+            if (arr[n + x].keyword == "CHECK")
+              break;
+          }
+          if (x == 3)
+          {
+            n += 2;
+            continue;
+          }
+          x += n + 1;
+          int depth = arr[n].depth;
+          int basedepth = arr[x].depth;
+          for (; x < arr.Length; x++)
+          {
+            if (arr[x].depth < basedepth)
+              break;
+
+            if (builder.Length > 0)
+              builder.Append(" ");
+
+            while (depth < arr[x].depth)
+            {
+              builder.Append("(");
+              depth++;
+            }
+            while (depth > arr[x].depth)
+            {
+              builder.Append(")");
+              depth--;
+            }
+
+            if (String.IsNullOrEmpty(arr[x].quote) == false)
+              builder.Append(arr[x].quote[0]);
+            builder.Append(arr[x].value);
+            if (String.IsNullOrEmpty(arr[x].quote) == false)
+              builder.Append(arr[x].quote[1]);
+
+            if (arr[x].sep == true) break;
+          }
+          while (depth > arr[n].depth)
+          {
+            builder.Append(")");
+            depth--;
+          }
+          n = x;
+          _check.Add(builder.ToString());
+        }
+      }
+
+      builder.Length = 0;
       builder.Append(_origSql);
+
       builder.AppendLine(";");
       foreach (Index idx in _oldindexes)
       {
         builder.AppendFormat("{0};\r\n", idx.OriginalSql);
+      }
+
+      _triggers.Clear();
+      _oldtriggers.Clear();
+
+      using (DataTable tbl = _connection.GetSchema("Triggers", new string[] { Catalog, null, Name }))
+      {
+        foreach (DataRow row in tbl.Rows)
+        {
+          Trigger t = new Trigger(this, row);
+          _triggers.Add(t);
+          _oldtriggers.Add(((ICloneable)t).Clone() as Trigger);
+
+          builder.AppendFormat("{0};\r\n", t.OriginalSql);
+        }
       }
 
       _origSql = builder.ToString();
@@ -115,6 +214,18 @@ namespace SQLite.Designer.Design
     {
       _exists = true;
       ReloadDefinition();
+
+      foreach (Column c in Columns)
+        c.Committed();
+
+      foreach (ForeignKey key in ForeignKeys)
+        key.ClearDirty();
+
+      foreach (Index idx in Indexes)
+        idx.ClearDirty();
+
+      if (PrimaryKey != null)
+        PrimaryKey.ClearDirty();
     }
 
     [Browsable(false)]
@@ -141,6 +252,18 @@ namespace SQLite.Designer.Design
     }
 
     [Browsable(false)]
+    public List<string> Check
+    {
+      get { return _check; }
+    }
+
+    [Browsable(false)]
+    public override object Triggers
+    {
+      get { return _triggers; }
+    }
+
+    [Browsable(false)]
     public string OriginalSql
     {
       get { return _origSql; }
@@ -155,7 +278,7 @@ namespace SQLite.Designer.Design
     [Category("Storage")]
     [RefreshProperties(RefreshProperties.All)]
     [ParenthesizePropertyName(true)]
-    public string Name
+    public override string Name
     {
       get { return _name; }
       set
@@ -178,26 +301,11 @@ namespace SQLite.Designer.Design
     [Editor(typeof(CatalogTypeEditor), typeof(UITypeEditor))]
     [DefaultValue("main")]
     [RefreshProperties(RefreshProperties.All)]
-    public string Catalog
+    public override string Catalog
     {
       get { return _catalog; }
-      //set
-      //{
-      //  string catalogs = "";
-      //  using (DataTable table = _connection.GetSchema("Catalogs"))
-      //  {
-      //    foreach (DataRow row in table.Rows)
-      //    {
-      //      catalogs += (row[0].ToString() + ",");
-      //    }
-      //  }
-
-      //  if (catalogs.IndexOf(value + ",", StringComparison.OrdinalIgnoreCase) == -1)
-      //    throw new ArgumentOutOfRangeException("Unrecognized catalog!");
-
-      //  _catalog = value;
-      //}
     }
+
     [Category("Storage")]
     public string Database
     {
@@ -220,36 +328,81 @@ namespace SQLite.Designer.Design
         Guid g = Guid.NewGuid();
         altName = String.Format("{0}_{1}", Name, g.ToString("N"));
 
-        foreach (Index idx in _oldindexes)
+        if (_oldindexes.Count > 0)
         {
-          builder.AppendFormat("DROP INDEX [{0}].[{1}];\r\n", _catalog, idx.Name);
+          builder.Append("-- Drop previous indexes on the table\r\n");
+          foreach (Index idx in _oldindexes)
+          {
+            builder.AppendFormat("DROP INDEX [{0}].[{1}];\r\n", _catalog, idx.Name);
+          }
+          builder.AppendLine();
         }
-        builder.AppendFormat("ALTER TABLE [{0}].[{1}] RENAME TO [{2}];\r\n", _catalog, _oldname, altName);
+
+        if (_oldtriggers.Count > 0)
+        {
+          builder.Append("-- Drop previous triggers on the table\r\n");
+          foreach (Trigger trig in _oldtriggers)
+          {
+            builder.AppendFormat("DROP TRIGGER [{0}].[{1}];\r\n", _catalog, trig.Name);
+          }
+          builder.AppendLine();
+        }
+
+        builder.Append("-- Rename the old table\r\n");
+        builder.AppendFormat("ALTER TABLE [{0}].[{1}] RENAME TO [{2}];\r\n\r\n", _catalog, _oldname, altName);
       }
 
+      builder.Append("-- Create the new table\r\n");
       builder.AppendFormat("CREATE TABLE [{0}].[{1}] (\r\n", _catalog, Name);
       string separator = "    ";
 
       foreach (Column c in Columns)
       {
-        builder.Append(separator);
-        c.WriteSql(builder);
-        separator = ",\r\n    ";
+        if (String.IsNullOrEmpty(c.ColumnName) == false)
+        {
+          builder.Append(separator);
+          c.WriteSql(builder);
+          separator = ",\r\n    ";
+        }
       }
 
       if (_key.Columns.Count > 1)
       {
         string innersep = "";
-        builder.AppendFormat("{0}CONSTRAINT PK_{1} PRIMARY KEY (", separator, Name);
+        builder.AppendFormat("{0}CONSTRAINT [PK_{1}] PRIMARY KEY (", separator, Name);
         foreach (IndexColumn c in _key.Columns)
         {
-          builder.AppendFormat("{0}[{1}]{2}", innersep, c.Column, (c.SortMode == ColumnSortMode.Ascending) ? "" : " DESC");
+          builder.AppendFormat("{0}[{1}]", innersep, c.Column);
+          if (String.IsNullOrEmpty(c.Collate) == false && String.Compare(c.Collate, "BINARY", StringComparison.OrdinalIgnoreCase) != 0)
+            builder.AppendFormat(" COLLATE {0}", c.Collate.ToUpperInvariant());
+
+          if (c.SortMode != ColumnSortMode.Ascending)
+            builder.AppendFormat(" DESC");
+
           innersep = ", ";
         }
         builder.Append(")");
 
         if (_key.Conflict != ConflictEnum.Abort)
-          builder.AppendFormat(" ON CONFLICT {0}", _key.Conflict.ToString());
+          builder.AppendFormat(" ON CONFLICT {0}", _key.Conflict.ToString().ToUpperInvariant());
+      }
+
+      for (int n = 0; n < Check.Count; n++)
+      {
+        string check = Check[n];
+
+        if (String.IsNullOrEmpty(check) == true) continue;
+        SimpleTokenizer.StringParts[] arr = SimpleTokenizer.BreakString(check);
+        for (int x = 0; x < arr.Length; x++)
+        {
+          if (arr[x].depth == 0)
+          {
+            check = String.Format("({0})", check);
+            break;
+          }
+        }
+        builder.Append(separator);
+        builder.AppendFormat("CONSTRAINT [CK_{0}_{1}] CHECK {2}", Name, n + 1, check);
       }
 
       foreach (ForeignKey fkey in ForeignKeys)
@@ -264,6 +417,7 @@ namespace SQLite.Designer.Design
       if (altName != null)
       {
         separator = "";
+        builder.Append("\r\n-- Copy the contents of the old table into the new table\r\n");
         builder.AppendFormat("INSERT INTO [{0}].[{1}] (", _catalog, Name);
         foreach (Column c in Columns)
         {
@@ -273,7 +427,7 @@ namespace SQLite.Designer.Design
             separator = ", ";
           }
         }
-        builder.Append(") SELECT ");
+        builder.Append(")\r\n  SELECT ");
         separator = "";
         foreach (Column c in Columns)
         {
@@ -283,29 +437,43 @@ namespace SQLite.Designer.Design
             separator = ", ";
           }
         }
-        builder.AppendFormat(" FROM [{0}].[{1}];\r\n", _catalog, altName);
+        builder.AppendFormat("\r\n  FROM [{0}].[{1}];\r\n\r\n", _catalog, altName);
 
+        builder.Append("-- Drop the old table\r\n");
         builder.AppendFormat("DROP TABLE [{0}].[{1}];\r\n", _catalog, altName);
       }
-      separator = "";
-      foreach (Index idx in _indexes)
+
+      separator = "\r\n";
+      if (_indexes.Count > 0)
       {
-        builder.Append(separator);
-        idx.WriteSql(builder);
-        separator = "\r\n";
+        builder.Append("\r\n-- Create the new indexes");
+        foreach (Index idx in _indexes)
+        {
+          builder.Append(separator);
+          idx.WriteSql(builder);
+        }
+        builder.AppendLine();
+      }
+
+      if (_triggers.Count > 0)
+      {
+        builder.Append("\r\n-- Create the new triggers");
+        foreach (Trigger trig in _triggers)
+        {
+          builder.Append(separator);
+          trig.WriteSql(builder);
+          separator = "\r\n";
+        }
+        builder.AppendLine();
       }
 
       return builder.ToString();
     }
 
-    #region IHaveConnection Members
-
-    public DbConnection GetConnection()
+    public override DbConnection GetConnection()
     {
       return _connection;
     }
-
-    #endregion
 
     #region ICustomTypeDescriptor Members
 
@@ -383,6 +551,21 @@ namespace SQLite.Designer.Design
     string CatalogScope { get; }
     [Browsable(false)]
     string TableScope { get; }
+  }
+
+  internal class CollationTypeEditor : ObjectSelectorEditor
+  {
+    public override UITypeEditorEditStyle GetEditStyle(ITypeDescriptorContext context)
+    {
+      return UITypeEditorEditStyle.DropDown;
+    }
+
+    protected override void FillTreeWithData(Selector selector, ITypeDescriptorContext context, IServiceProvider provider)
+    {
+      base.FillTreeWithData(selector, context, provider);
+      selector.AddNode("BINARY", "BINARY", null);
+      selector.AddNode("NOCASE", "NOCASE", null);
+    }
   }
 
   internal class CatalogTypeEditor : ObjectSelectorEditor
